src/api_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
websockets==12.0
pydantic-settings==2.1.0
asyncpg==0.29.0
